// vercel-daily/lib/types/news-api.d.ts
export interface components {
    schemas: {
        Author: {
            /** @example Rhys Sullivan */
            name?: string;
            /** @example  */
            avatar?: string;
        };
        ContentBlock: {
            /** @enum {string} */
            type: "paragraph";
            text: string;
        } | {
            /** @enum {string} */
            type: "heading";
            /** @enum {integer} */
            level: 2 | 3;
            text: string;
        } | {
            /** @enum {string} */
            type: "blockquote";
            text: string;
        } | {
            /** @enum {string} */
            type: "unordered-list";
            items: string[];
        } | {
            /** @enum {string} */
            type: "ordered-list";
            items: string[];
        } | {
            /** @enum {string} */
            type: "image";
            /** Format: uri */
            src: string;
            alt: string;
            caption?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}